# swagger.yaml
swagger: "2.0"
info:
  description: "This is the specification for the API of the Frontend-Service of the MemeIt-Project for maue-wise2021. The frontend and this API is written by Talha Dagli, John Einicke and Yoel Berendsohn."
  version: "2.1.0"
  title: "MemeIt-Frontend-Service"
host: "host.frontend.io"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /users/logout:
    get:
      tags:
      - "User"
      summary: "Log out user"
      description: "Log out the user. Requires the sending of the cookie containing the authToken created by Log in."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Logout operation"
          schema:
            properties:
             message:
              type: string
            example:
              message: "Successfully logged out"
        401:
          description: "Not logged in"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Not logged in"
  /users/login:
    post:
      tags:
      - "User"
      summary: "Log in user"
      description: "Takes an username and a password and tries to log the user in."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "credentials"
        description: "Username and password"
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          example:
            email: "HansRolf@rolf.de"
            password: "password"
      responses:
        200:
          description: "Successful login operation. Returns cookie with token."
          schema:
            properties:
             message:
              type: string
            example:
              message: "Successfully logged in"
        401:
          description: "Wrong email or password"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Wrong email or password"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        400:
          description: "Invalid Request, when parameters are missing"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Invalid Request"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"
  /users/register:
    post:
      tags:
      - "User"
      summary: "Register in user"
      description: "Takes an username, email address and a password and registers the user."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Credentials"
        description: "Username, password and email"
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
          example:
            email: "email@adress.com"
            name: "Hans"
            password: "Rolf"
      responses:
        201:
          description: "Successful register operation."
          schema:
            properties:
             message:
              type: string
            example:
              message: "Successfully registered"
        409:
          description: "User with this email already exists"
          schema:
            properties:
              message:
                type: string
            example:
              message: "User with this email already exist"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        400:
          description: "Invalid Request, when parameters are missing"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Invalid Request"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"
  /users/delete:
    post:
      tags:
      - "User"
      summary: "Delete in user"
      description: "Requires the sending of the cookie containing the authToken created by Log in. Deletes the currently logged in user."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Successfully deleted user"
          schema:
            properties:
             message:
              type: string
            example:
              message: "Successfully deleted"
        401:
          description: "Not logged in"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Not logged in"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"
  /users/favorites:
    get:
      tags:
      - "User"
      summary: "Get favorites of user"
      description: "Requires the sending of the cookie containing the authToken created by Log in. Gets the favorites of the currently logged in user."
      produces:
      - "application/json"
      responses:
        200:
          description: "Favorites of the user"
          schema:
            properties:
              response:
                type: array
            example: 
              - id: 0
                memeId: string
              - id: 1
                memeId: string
        401:
          description: "Not logged in"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Not logged in"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"
    post:
      tags:
      - "User"
      summary: "Favorize meme of user"
      description: "Requires the sending of the cookie containing the authToken created by Log in. Favorize a meme for the currently logged in user."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Body"
        description: "Username, password and email"
        required: true
        schema:
          properties:
            memeId:
              type: string
          example:
            memeId: "string"
      responses:
        200:
          description: "Favorite of the user"
          schema:
            properties:
              response:
                type: array
            example: 
              - id: 0
                memeId: string
                user:
                  id: 1
        400:
          description: "Invalid Request, when parameters are missing"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Invalid Request"
        401:
          description: "Not logged in"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Not logged in"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"    
  /users/favorites/{memeid}:
    delete:
      tags:
      - "User"
      summary: "Deletes a favorite of user"
      description: "Requires the sending of the cookie containing the authToken created by Log in. Deletes a favorite of the currently logged in user."
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "memeid"
        required: true
        type:
          string
      responses:
        200:
          description: "Favorites of the user"
        401:
          description: "Not logged in"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Not logged in"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              userServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              userServiceResponse: "Message from User Service"
        408:
          description: "User Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from User Service timed out"         
...